pyproject.toml
[project]
name = "topic-mapped-text-embeddings"
version = "0.1.0"
description = "Topic Mapped Text Embeddings for Effective Retrieval"
requires-python = ">=3.9,<4.0"

dependencies = [
    "langchain-community>=0.3.20",
    "nltk>=3.9.1",
    "pandas",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "tokenizer>=3.4.4",
    "tqdm",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "faiss-cpu>=1.7.4",
    "ftfy",
    "regex",
    "tiktoken",
    "openai-clip>=1.0.1",
    "setuptools>=78.1.0",
    "torchvision>=0.21.0",
    "sentence-transformers>=4.0.1",
    "streamlit>=1.44.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."] 
